Added further tests
