+++ force_wfmanager/wfmanager_setup_task.py
@@ -1,30 +1,30 @@
 import logging
 import textwrap
+import subprocess
+import tempfile
+import os
+from concurrent.futures import ThreadPoolExecutor
 
-from envisage.ui.tasks.tasks_application import TasksApplication
-from force_bdss.io.workflow_reader import InvalidFileException, WorkflowReader
-from force_bdss.io.workflow_writer import WorkflowWriter
-from force_bdss.ui_hooks.base_ui_hooks_manager import BaseUIHooksManager
-
-from pyface.api import ImageResource
-from pyface.constant import OK
-from pyface.file_dialog import FileDialog
-from pyface.message_dialog import error, information
+from pyface.api import (ImageResource, confirm, YES, GUI, error, information,
+                        OK, FileDialog)
 from pyface.tasks.action.api import SMenuBar, SMenu, TaskAction, SToolBar
-from pyface.tasks.action.schema_addition import SchemaAddition
 from pyface.tasks.api import Task, TaskLayout, PaneItem
 
-from traits.api import Instance, on_trait_change, List, Bool
+from traits.api import Instance, on_trait_change, List, Bool, Unicode, File
 
-from force_bdss.api import IFactoryRegistryPlugin, Workflow
+from force_bdss.api import (BaseExtensionPlugin, BaseUIHooksManager,
+                            IFactoryRegistryPlugin, InvalidFileException,
+                            MCOProgressEvent, MCOStartEvent, Workflow,
+                            WorkflowReader, WorkflowWriter)
 
 from force_wfmanager.central_pane.analysis_model import AnalysisModel
 from force_wfmanager.central_pane.setup_pane import SetupPane
+from force_wfmanager.plugin_dialog import PluginDialog
 from force_wfmanager.left_side_pane.tree_pane import TreePane
 from force_wfmanager.TaskToggleGroupAccelerator import (
     TaskToggleGroupAccelerator
 )
-from traits.trait_types import File
+from force_wfmanager.server.zmq_server import ZMQServer
 
 log = logging.getLogger(__name__)
 
@@ -35,6 +35,10 @@ class WfManagerSetupTask(Task):
 
     #: Workflow model.
     workflow_m = Instance(Workflow, allow_none=False)
+
+    #: Analysis model. Contains the results that are displayed in the plot
+    #: and table
+    analysis_m = Instance(AnalysisModel, allow_none=False)
 
     #: Registry of the available factories
     factory_registry = Instance(IFactoryRegistryPlugin)
@@ -57,6 +61,19 @@ class WfManagerSetupTask(Task):
     #: Are the saving and loading menu/toolbar buttons active
     save_load_enabled = Bool(True)
 
+    #: Is there a bdss computation running
+    computation_running = Bool(False)
+
+    #: The thread pool executor to spawn the BDSS CLI process.
+    executor = Instance(ThreadPoolExecutor)
+
+    #: Path to spawn for the BDSS CLI executable.
+    #: This will go to some global configuration option later.
+    bdss_executable_path = Unicode("force_bdss")
+
+    #: ZeroMQ Server to receive information from the running BDSS
+    zmq_server = Instance(ZMQServer)
+
     #: A list of UI hooks managers. These hold plugin injected "hook managers",
     #: classes with methods that are called when some operation is performed
     #: by the UI
@@ -64,9 +81,22 @@ class WfManagerSetupTask(Task):
 
     task_group = Instance(TaskToggleGroupAccelerator)
 
+    #: Results Task
+    results_task = Instance(Task)
+
     def __init__(self, factory_registry, *args, **kwargs):
         self.factory_registry = factory_registry
         super(WfManagerSetupTask, self).__init__(*args, **kwargs)
+
+    # ZMQ Setup
+
+    def initialized(self):
+        self.zmq_server.start()
+
+    def prepare_destroy(self):
+        self.zmq_server.stop()
+
+    # Task Defaults and Initialisers
 
     def _menu_bar_default(self):
         """A menu bar with functions relevant to the Setup task.
@@ -176,6 +206,12 @@ class WfManagerSetupTask(Task):
             )
         ]
 
+    def _default_layout_default(self):
+        """ Defines the default layout of the task window """
+        return TaskLayout(
+            left=PaneItem('force_wfmanager.tree_pane'),
+        )
+
     def create_central_pane(self):
         """ Creates the central pane which contains the analysis part
         (pareto front and output KPI values)
@@ -188,12 +224,6 @@ class WfManagerSetupTask(Task):
 
     # Default initializers
 
-    def _default_layout_default(self):
-        """ Defines the default layout of the task window """
-        return TaskLayout(
-            left=PaneItem('force_wfmanager.tree_pane'),
-        )
-
     def _side_pane_default(self):
         return TreePane(
             factory_registry=self.factory_registry,
@@ -202,6 +232,9 @@ class WfManagerSetupTask(Task):
 
     def _workflow_m_default(self):
         return Workflow()
+
+    def _analysis_m_default(self):
+        return AnalysisModel()
 
     def _ui_hooks_managers_default(self):
         hooks_factories = self.factory_registry.ui_hooks_factories
@@ -219,7 +252,27 @@ class WfManagerSetupTask(Task):
                 )
         return managers
 
+    def _executor_default(self):
+        return ThreadPoolExecutor(max_workers=1)
+
+    def _zmq_server_default(self):
+        return ZMQServer(
+            on_event_callback=self._server_event_callback,
+            on_error_callback=self._server_error_callback
+        )
+
     # Workflow Methods
+
+    def open_workflow(self):
+        """ Shows a dialog to open a workflow file """
+        dialog = FileDialog(
+            action="open",
+            wildcard='JSON files (*.json)|*.json|'
+        )
+        result = dialog.open()
+        f_name = dialog.path
+        if result is OK:
+            self.open_workflow_file(f_name)
 
     def open_workflow_file(self, f_name):
         """ Opens a workflow from the specified file name"""
@@ -317,17 +370,6 @@ class WfManagerSetupTask(Task):
         else:
             return True
 
-    def open_workflow(self):
-        """ Shows a dialog to open a workflow file """
-        dialog = FileDialog(
-            action="open",
-            wildcard='JSON files (*.json)|*.json|'
-        )
-        result = dialog.open()
-        f_name = dialog.path
-        if result is OK:
-            self.open_workflow_file(f_name)
-
     def open_about(self):
         information(
             None,
@@ -343,45 +385,217 @@ class WfManagerSetupTask(Task):
             "About WorkflowManager"
         )
 
-    # Sync Handlers
+    # BDSS Interaction
 
-    # Inbound handlers
+    def run_bdss(self):
+        """ Run the BDSS computation """
+
+        # Confirm we want to run a calculation
+        if len(self.analysis_m.evaluation_steps) != 0:
+            result = confirm(
+                None,
+                "Are you sure you want to run the computation and "
+                "empty the result table?")
+            if result is not YES:
+                return
+
+        # Flag that a computation is running (passed to application-level)
+        self.computation_running = True
+        try:
+            for hook_manager in self.ui_hooks_managers:
+                try:
+                    hook_manager.before_execution(self)
+                except Exception:
+                    log.exception(
+                        "Failed before_execution hook "
+                        "for hook manager {}".format(
+                            hook_manager.__class__.__name__)
+                    )
+
+            # Creates a temporary file containing the workflow
+            tmpfile_path = tempfile.mktemp()
+            with open(tmpfile_path, 'w') as output:
+                WorkflowWriter().write(self.workflow_m, output)
+
+            # Clear the analysis model before attempting to run
+            self.analysis_m.clear()
+
+            future = self.executor.submit(self._execute_bdss, tmpfile_path)
+            future.add_done_callback(self._execution_done_callback)
+        except Exception as e:
+            logging.exception("Unable to run BDSS.")
+            error(None,
+                  "Unable to run BDSS: {}".format(e),
+                  'Error when running BDSS'
+                  )
+            self.computation_running = False
+
+    def _execute_bdss(self, workflow_path):
+        """Secondary thread executor routine.
+        This executes the BDSS and wait for its completion.
+        """
+        try:
+            subprocess.check_call([self.bdss_executable_path,
+                                   workflow_path])
+        except OSError as e:
+            log.exception("Error while executing force_bdss executable. "
+                          " Is force_bdss in your path?")
+            self._clean_tmp_workflow(workflow_path, silent=True)
+            raise e
+        except subprocess.CalledProcessError as e:
+            # Ignore any error of execution.
+            log.exception("force_bdss returned a "
+                          "non-zero value after execution")
+            self._clean_tmp_workflow(workflow_path, silent=True)
+            raise e
+        except Exception as e:
+            log.exception("Unknown exception occurred "
+                          "while invoking force bdss executable.")
+            self._clean_tmp_workflow(workflow_path, silent=True)
+            raise e
+
+        self._clean_tmp_workflow(workflow_path)
+
+    def _clean_tmp_workflow(self, workflow_path, silent=False):
+        """Removes the temporary file for the workflow.
+
+        Parameters
+        ----------
+        workflow_path: str
+            The path of the workflow
+        silent: bool
+            If true, any exception encountered will be discarded (but logged).
+            If false, the exception will be re-raised
+        """
+        try:
+            os.remove(workflow_path)
+        except OSError as e:
+            # Ignore deletion errors, in case the file magically
+            # vanished in the meantime
+            log.exception("Unable to delete temporary "
+                          "workflow file at {}".format(workflow_path))
+            if not silent:
+                raise e
+
+    def _execution_done_callback(self, future):
+        """Secondary thread code.
+        Called when the execution is completed.
+        """
+        exc = future.exception()
+        GUI.invoke_later(self._bdss_done, exc)
+
+    def _bdss_done(self, exception):
+        """Called in the main thread when the execution is completed.
+
+        Parameters
+        ----------
+        exception: Exception or None
+            If the execution raised an exception of any sort.
+        """
+
+        for hook_manager in self.ui_hooks_managers:
+            try:
+                hook_manager.after_execution(self)
+            except Exception:
+                log.exception(
+                    "Failed after_execution hook "
+                    "for hook manager {}".format(
+                        hook_manager.__class__.__name__)
+                )
+
+        self.computation_running = False
+
+        if exception is not None:
+            error(
+                None,
+                'Execution of BDSS failed. \n\n{}'.format(
+                    str(exception)),
+                'Error when running BDSS'
+            )
+
+    # Plugin Status
+
+    def open_plugins(self):
+        plugins = [plugin
+                   for plugin in self.window.application.plugin_manager
+                   if isinstance(plugin, BaseExtensionPlugin)]
+
+        # Plugins guaranteed to have an id, so sort by that if name is not set
+        plugins.sort(key=lambda s: s.name
+                     if s.name not in ('', None) else s.id)
+        dlg = PluginDialog(plugins)
+        dlg.edit_traits()
+
+    # Handling of BDSS events via ZMQ server
+
+    def _server_event_callback(self, event):
+        """Callback that is called by the server thread
+        when a new event is received. This method is
+        executed by the server thread.
+        """
+        GUI.invoke_later(self._server_event_mainthread, event)
+
+    def _server_error_callback(self, error_type, error_message):
+        """Callback in case of server error. Invoked by the secondary thread"""
+        if error_type == ZMQServer.ERROR_TYPE_CRITICAL:
+            GUI.invoke_later(self._show_error_dialog, error_message)
+
+    def _server_event_mainthread(self, event):
+        """Invoked by the main thread.
+        Handles the event received by the server, dispatching its
+        action appropriately according to the type"""
+        if isinstance(event, MCOStartEvent):
+            self.analysis_m.clear()
+            value_names = list(event.parameter_names)
+            for kpi_name in event.kpi_names:
+                value_names.extend([kpi_name, kpi_name + " weight"])
+            self.analysis_m.value_names = tuple(value_names)
+        elif isinstance(event, MCOProgressEvent):
+            data = [dv.value for dv in event.optimal_point]
+            for kpi, weight in zip(event.optimal_kpis, event.weights):
+                data.extend([kpi.value, weight])
+
+            data = tuple(map(float, data))
+            self.analysis_m.add_evaluation_step(data)
+
+    # Error Display
+
+    def _show_error_dialog(self, message):
+        """Shows an error dialog to the user with a given message"""
+        error(None, message, "Server error")
+
+    # Synchronization with side pane (Tree Pane)
 
     @on_trait_change('side_pane.run_enabled')
     def set_toolbar_run_btn_state(self):
         self.run_enabled = self.side_pane.run_enabled
 
-    @on_trait_change('window.application.computation_running')
-    def update_side_pane_status(self):
-        if self.window is not None:
-            self.side_pane.ui_enabled = not self.window.application.computation_running
-            self.save_load_enabled = not self.window.application.computation_running
+    @on_trait_change('workflow_m')
+    def update_side_pane_model(self):
+        self.side_pane.workflow_m = self.workflow_m
 
-    # Outbound Handlers
-
-    @on_trait_change('run_enabled')
-    def update_wfmanager_run_enabled(self):
+    @on_trait_change('computation_running')
+    def update_pane_active_status(self):
         if self.window is not None:
-            self.window.application.run_enabled = self.run_enabled
+            self.side_pane.ui_enabled = not self.computation_running
+            self.save_load_enabled = not self.computation_running
+
+    # Synchronization with Window
+
+    @on_trait_change('window.tasks')
+    def get_results_task(self):
+        if self.window is not None:
+            for task in self.window.tasks:
+                if task.name == "Results":
+                    self.results_task = task
+                    self.results_task.run_enabled = self.run_enabled
 
     # Menu/Toolbar Methods
 
     def switch_task(self):
         """Switches to the results task and verifies startup setting are
         correct for toolbars/menus etc."""
-        results_task = self.window.get_task(
-            'force_wfmanager.wfmanager_results_task'
-        )
-        results_task.run_enabled = self.run_enabled
-        self.window.activate_task(results_task)
+        self.window.activate_task(self.results_task)
 
     def exit(self):
         self.window.application.exit()
-
-    def open_plugins(self):
-        self.window.application.open_plugins()
-
-    def run_bdss(self):
-        task = self.window.get_task('force_wfmanager.wfmanager_results_task')
-        task.initialized()
-        task.run_bdss()
